[GitHubEnterprise-3.6] Update to 1.1.4 (hash: 04d3e6d3487034c5ce6388f09ae00d4b) from 1.1.4 (hash: 0a642d59cf98ed08f2fe97f53aa6a330) Detected Schema changes:
├─┬Paths
│ ├─┬/search/issues
│ │ └─┬GET
│ │   └──[M] description (33421:20)
│ ├─┬/applications/{client_id}/token/scoped
│ │ └─┬POST
│ │   ├──[M] description (2363:20)
│ │   └─┬Requestbody
│ │     └─┬application/json
│ │       └─┬Schema
│ │         ├─┬repositories
│ │         │ └──[M] description (2404:32)
│ │         ├─┬repository_ids
│ │         │ └──[M] description (2413:32)
│ │         ├─┬target
│ │         │ └──[M] description (2392:32)
│ │         └─┬target_id
│ │           └──[M] description (2398:32)
│ ├─┬/user/installations
│ │ └─┬GET
│ │   └──[M] description (36516:20)
│ └─┬/user/installations/{installation_id}/repositories
│   └─┬GET
│     └──[M] description (36571:20)
└─┬Components
  └─┬app-permissions
    └──[M] description (39834:20)

Date: 06/20/23 | Commit: New: etc/specs/GitHubEnterprise-3.6/previous.spec.yaml, Original: etc/specs/GitHubEnterprise-3.6/current.spec.yaml
Document Element | Total Changes | Breaking Changes
paths            | 8             | 0
components       | 1             | 0
INFO: Total Changes: 9
