+++ etc/specs/GitHubEnterprise-3.6/current.spec.yaml
@@ -2360,10 +2360,10 @@ paths:
   "/applications/{client_id}/token/scoped":
     post:
       summary: Create a scoped access token
-      description: Use a non-scoped user-to-server access token to create a repository
-        scoped and/or permission scoped user-to-server access token. You can specify
-        which repositories the token can access and which permissions are granted
-        to the token. You must use [Basic Authentication](https://docs.github.com/enterprise-server@3.6/rest/overview/other-authentication-methods#basic-authentication)
+      description: Use a non-scoped user access token to create a repository scoped
+        and/or permission scoped user access token. You can specify which repositories
+        the token can access and which permissions are granted to the token. You must
+        use [Basic Authentication](https://docs.github.com/enterprise-server@3.6/rest/overview/other-authentication-methods#basic-authentication)
         when accessing this endpoint, using the `client_id` and `client_secret` of
         the GitHub App as the username and password. Invalid tokens will return `404
         NOT FOUND`.
@@ -2389,20 +2389,20 @@ paths:
                   examples:
                   - e72e16c7e42f292c6912e7710c838347ae178b4a
                 target:
-                  description: The name of the user or organization to scope the user-to-server
+                  description: The name of the user or organization to scope the user
                     access token to. **Required** unless `target_id` is specified.
                   type: string
                   examples:
                   - octocat
                 target_id:
-                  description: The ID of the user or organization to scope the user-to-server
+                  description: The ID of the user or organization to scope the user
                     access token to. **Required** unless `target` is specified.
                   type: integer
                   examples:
                   - 1
                 repositories:
-                  description: The list of repository names to scope the user-to-server
-                    access token to. `repositories` may not be specified if `repository_ids`
+                  description: The list of repository names to scope the user access
+                    token to. `repositories` may not be specified if `repository_ids`
                     is specified.
                   type: array
                   items:
@@ -2410,8 +2410,8 @@ paths:
                     examples:
                     - rails
                 repository_ids:
-                  description: The list of repository IDs to scope the user-to-server
-                    access token to. `repository_ids` may not be specified if `repositories`
+                  description: The list of repository IDs to scope the user access
+                    token to. `repository_ids` may not be specified if `repositories`
                     is specified.
                   type: array
                   items:
@@ -33430,7 +33430,7 @@ paths:
 
         This query searches for the keyword `windows`, within any open issue that is labeled as `bug`. The search runs across repositories whose primary language is Python. The results are sorted by creation date in ascending order, which means the oldest issues appear first in the search results.
 
-        **Note:** For [user-to-server](https://docs.github.com/enterprise-server@3.6/developers/apps/identifying-and-authorizing-users-for-github-apps#user-to-server-requests) GitHub App requests, you can't retrieve a combination of issues and pull requests in a single query. Requests that don't include the `is:issue` or `is:pull-request` qualifier will receive an HTTP `422 Unprocessable Entity` response. To get results for both issues and pull requests, you must send separate queries for issues and pull requests. For more information about the `is` qualifier, see "[Searching only issues or pull requests](https://docs.github.com/enterprise-server@3.6/github/searching-for-information-on-github/searching-issues-and-pull-requests#search-only-issues-or-pull-requests)."
+        **Note:** For requests made by GitHub Apps with a user access token, you can't retrieve a combination of issues and pull requests in a single query. Requests that don't include the `is:issue` or `is:pull-request` qualifier will receive an HTTP `422 Unprocessable Entity` response. To get results for both issues and pull requests, you must send separate queries for issues and pull requests. For more information about the `is` qualifier, see "[Searching only issues or pull requests](https://docs.github.com/enterprise-server@3.6/github/searching-for-information-on-github/searching-issues-and-pull-requests#search-only-issues-or-pull-requests)."
       tags:
       - search
       operationId: search/issues-and-pull-requests
@@ -36516,7 +36516,7 @@ paths:
       description: |-
         Lists installations of your GitHub App that the authenticated user has explicit permission (`:read`, `:write`, or `:admin`) to access.
 
-        You must use a [user-to-server OAuth access token](https://docs.github.com/enterprise-server@3.6/apps/building-github-apps/identifying-and-authorizing-users-for-github-apps/#identifying-users-on-your-site), created for a user who has authorized your GitHub App, to access this endpoint.
+        You must use a [user access token](https://docs.github.com/enterprise-server@3.6/apps/creating-github-apps/authenticating-with-a-github-app/generating-a-user-access-token-for-a-github-app), created for a user who has authorized your GitHub App, to access this endpoint.
 
         The authenticated user has explicit permission to access repositories they own, repositories where they are a collaborator, and repositories that they can access through an organization membership.
 
@@ -36573,7 +36573,7 @@ paths:
 
         The authenticated user has explicit permission to access repositories they own, repositories where they are a collaborator, and repositories that they can access through an organization membership.
 
-        You must use a [user-to-server OAuth access token](https://docs.github.com/enterprise-server@3.6/apps/building-github-apps/identifying-and-authorizing-users-for-github-apps/#identifying-users-on-your-site), created for a user who has authorized your GitHub App, to access this endpoint.
+        You must use a [user access token](https://docs.github.com/enterprise-server@3.6/apps/creating-github-apps/authenticating-with-a-github-app/generating-a-user-access-token-for-a-github-app), created for a user who has authorized your GitHub App, to access this endpoint.
 
         The access the user has to each repository is included in the hash under the `permissions` key.
       tags:
@@ -39831,7 +39831,7 @@ components:
     app-permissions:
       title: App Permissions
       type: object
-      description: The permissions granted to the user-to-server access token.
+      description: The permissions granted to the user access token.
       properties:
         actions:
           type: string
